refactor: simplify SR11 policy seam (reuse + de-dup, behaviour-neutral)
/simplify (be-review) — 4 cleanup agents, deduped:
- surface-map: pass `appOnError` straight to the entries `buildSurfaceClient` — its
  optional 3rd `origin` param is already assignable to the origin-free `OnClientError`,
  so the forwarding wrapper stripped nothing (and `undefined` still flows through, keeping
  the fail-fast reachable).
- wire.ts `interpretClientError`: extracted `requireOrigin` (collapses the two identical
  copy-paste origin-guard throws) and reused the already-computed `originEnc` instead of
  recomputing `encodeHostKey(origin.key)` twice in the scopedSub arm.
- hoisted the string-aware comment-stripper into the shared `architectureGuardSources.testlib`
  as `stripComments` — it was duplicated as `blankComments` here and (a weaker regex form)
  `stripComments` in `standingSubscriptionOwnership.test.ts`; both guard tests now share the
  one robust helper.
Altitude confirmed right depth (the terminalKeys stream's inline policy is a defensible
health carve-out, not a bandaid — left as-is).

SR11-POLICY.
